[WRFE-70](refactor): 태그 커스텀 훅으로 관리 변경
